Extract shared CarrierTermsView from UPSTermsView
Rename UPSTOS/ to CarrierTOS/ and introduce:
- CarrierTermsViewModel protocol for carrier-specific ToS view models
- CarrierTermsView generic view with @ViewBuilder checkbox injection
- LinkedCheckboxToggle reusable helper for checkbox-with-link pattern

UPSTermsViewModel now conforms to CarrierTermsViewModel. The old
UPSTermsView is replaced by CarrierTermsView at the call site.
