+++ WooCommerce/Classes/ViewRelated/Orders/Order Details/Shipping Labels/WooShipping Create Shipping Labels/WooShippingCreateLabelsView.swift
@@ -125,12 +125,28 @@ struct WooShippingCreateLabelsView: View {
             }
             .sheet(isPresented: $viewModel.shouldShowUPSTermsAndConditions) {
                 if let termsViewModel = viewModel.upsTermsViewModel {
-                    UPSTermsView(viewModel: termsViewModel) {
+                    CarrierTermsView(viewModel: termsViewModel, checkboxes: { vm in
+                        LinkedCheckboxToggle(isOn: Binding(get: { vm.isTOSAccepted },
+                                                           set: { vm.isTOSAccepted = $0 }),
+                                             labelFormat: UPSTermsViewModel.Localization.checkbox1,
+                                             linkText: UPSTermsViewModel.Localization.termsOfService,
+                                             linkURL: UPSTermsViewModel.Links.termsOfService)
+                        LinkedCheckboxToggle(isOn: Binding(get: { vm.isProhibitedItemsAccepted },
+                                                           set: { vm.isProhibitedItemsAccepted = $0 }),
+                                             labelFormat: UPSTermsViewModel.Localization.checkbox2,
+                                             linkText: UPSTermsViewModel.Localization.prohibitedItems,
+                                             linkURL: UPSTermsViewModel.Links.prohibitedItems)
+                        LinkedCheckboxToggle(isOn: Binding(get: { vm.isTechnologyAgreementAccepted },
+                                                           set: { vm.isTechnologyAgreementAccepted = $0 }),
+                                             labelFormat: UPSTermsViewModel.Localization.checkbox3,
+                                             linkText: UPSTermsViewModel.Localization.technologyAgreement,
+                                             linkURL: UPSTermsViewModel.Links.techAgreement)
+                    }, onConfirmation: {
                         viewModel.shouldShowUPSTermsAndConditions = false
                         Task { @MainActor in
                             await viewModel.purchaseLabel(shouldRefreshPackageAndRate: true)
                         }
-                    }
+                    })
                     .presentationDetents([.large])
                 }
             }
